backend/requirements.txt
fastapi
uvicorn[standard]
mangum
mysql-connector-python
python-dotenv
python-multipart
passlib[bcrypt]==1.7.4
bcrypt==3.2.2
python-jose[cryptography]
pydantic[email]
boto3
psutil